// src/types/lightningType.ts
export interface LightningMeetup {
  lightningId: number;
  lightningName: string;
  latitude: number;
  longitude: number;
  lightningType: string;
  lightningTime: string;
  lightningDate: string;
  lightningCapacity: number;
  lightningParticipants: number;
  address: string;
  nickname: string;
  profileImagePath: string;
}